Implement MultiplexW and Pool
