+++ sdks/go/pkg/beam/runners/prism/internal/jobservices/management.go
@@ -92,8 +92,9 @@ func (s *Server) Prepare(ctx context.Context, req *jobpb.PrepareJobRequest) (_ *
 			cancelFn(err)
 			terminalOnceWrap()
 		},
-		Logger:           s.logger, // TODO substitute with a configured logger.
-		artifactEndpoint: s.Endpoint(),
+		Logger:             s.logger, // TODO substitute with a configured logger.
+		artifactEndpoint:   s.Endpoint(),
+		WorkerPoolEndpoint: s.WorkerPoolEndpoint,
 	}
 	// Stop the idle timer when a new job appears.
 	if idleTimer := s.idleTimer.Load(); idleTimer != nil {
