+++ sdks/go/pkg/beam/runners/prism/prism.go
@@ -27,6 +27,7 @@ import (
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal/jobservices"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal/web"
+	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/prism/internal/worker"
 	"github.com/apache/beam/sdks/v2/go/pkg/beam/runners/universal"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
@@ -62,6 +63,9 @@ type Options struct {
 	// Port the Job Management Server should start on.
 	Port int
 
+	// WorkerPoolEndpoint is the endpoint to connect with the worker pool service.
+	WorkerPoolEndpoint string
+
 	// The time prism will wait for new jobs before shuting itself down.
 	IdleShutdownTimeout time.Duration
 	// CancelFn allows Prism to terminate the program due to it's internal state, such as via the idle shutdown timeout.
@@ -73,6 +77,7 @@ type Options struct {
 // This call is non-blocking.
 func CreateJobServer(ctx context.Context, opts Options) (jobpb.JobServiceClient, error) {
 	s := jobservices.NewServer(opts.Port, internal.RunPipeline)
+	s.WorkerPoolEndpoint = opts.WorkerPoolEndpoint
 
 	if opts.IdleShutdownTimeout > 0 {
 		s.IdleShutdown(opts.IdleShutdownTimeout, opts.CancelFn)
@@ -89,4 +94,9 @@ func CreateJobServer(ctx context.Context, opts Options) (jobpb.JobServiceClient,
 // This call is blocking.
 func CreateWebServer(ctx context.Context, cli jobpb.JobServiceClient, opts Options) error {
 	return web.Initialize(ctx, opts.Port, cli)
+}
+
+// CreateWorkerPoolServer initializes the worker pool server that multiplexes worker.W gRPC requests.
+func CreateWorkerPoolServer(ctx context.Context) *grpc.Server {
+	return worker.NewMultiplexW()
 }
